refactor: convert login page to overlay component
