+++ components.d.ts
@@ -62,6 +62,7 @@ declare module 'vue' {
     FlashMessage: typeof import('./src/components/common/FlashMessage.vue')['default']
     KeyboardShortcutsDialog: typeof import('./src/components/common/KeyboardShortcutsDialog.vue')['default']
     KlippyStatusCard: typeof import('./src/components/common/KlippyStatusCard.vue')['default']
+    Login: typeof import('./src/components/common/Login.vue')['default']
     ManualProbeDialog: typeof import('./src/components/common/ManualProbeDialog.vue')['default']
     PeripheralsDialog: typeof import('./src/components/common/PeripheralsDialog.vue')['default']
     PromiseWrapper: typeof import('./src/components/common/PromiseWrapper.vue')['default']
@@ -122,8 +123,6 @@ declare module 'vue' {
     VOverlay: typeof import('vuetify/lib')['VOverlay']
     VProgressCircular: typeof import('vuetify/lib')['VProgressCircular']
     VProgressLinear: typeof import('vuetify/lib')['VProgressLinear']
-    VRadio: typeof import('vuetify/lib')['VRadio']
-    VRadioGroup: typeof import('vuetify/lib')['VRadioGroup']
     VRow: typeof import('vuetify/lib')['VRow']
     VSelect: typeof import('vuetify/lib')['VSelect']
     VSheet: typeof import('vuetify/lib')['VSheet']
@@ -132,11 +131,6 @@ declare module 'vue' {
     VSlider: typeof import('vuetify/lib')['VSlider']
     VSnackbar: typeof import('vuetify/lib')['VSnackbar']
     VSpacer: typeof import('vuetify/lib')['VSpacer']
-    VStepper: typeof import('vuetify/lib')['VStepper']
-    VStepperContent: typeof import('vuetify/lib')['VStepperContent']
-    VStepperHeader: typeof import('vuetify/lib')['VStepperHeader']
-    VStepperItems: typeof import('vuetify/lib')['VStepperItems']
-    VStepperStep: typeof import('vuetify/lib')['VStepperStep']
     VSubheader: typeof import('vuetify/lib')['VSubheader']
     VSwitch: typeof import('vuetify/lib')['VSwitch']
     VTab: typeof import('vuetify/lib')['VTab']
